Implementar gestión de imágenes del sitio (logo y banner)
- Crear nuevo componente SiteImageManagement para manejar la carga de logo y banner
- Actualizar AdminPage para incluir SiteImageManagement
- Implementar API route upload-image usando multer para manejar la carga de archivos
- Crear API route site-images para obtener las imágenes del sitio
- Actualizar la estructura de datos en Vercel KV para almacenar URLs de imágenes
- Ajustar tsconfig.json para manejar correctamente los tipos
- Reemplazar formidable con multer para la carga de archivos
- Implementar manejo de errores y validaciones en la carga de imágenes
