+++ package.json
@@ -20,6 +20,7 @@
     "framer-motion": "^10.13.1",
     "googleapis": "^122.0.0",
     "mercadopago": "^1.5.17",
+    "multer": "^1.4.5-lts.1",
     "next": "^13.4.12",
     "papaparse": "^5.4.1",
     "postcss": "^8.4.27",
@@ -31,6 +32,7 @@
     "uuid": "^9.0.0"
   },
   "devDependencies": {
+    "@types/multer": "^1.4.12",
     "@types/node": "^20.4.5",
     "@types/papaparse": "^5.3.7",
     "@types/react": "^18.2.17",
